pyproject.toml
[project]
name = "consilium"
version = "0.11.0"
description = "Multi-model deliberation CLI. 4 frontier LLMs debate with rotating challenger, then Claude judges."
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Terry Li", email = "terry.li.hm@gmail.com" }
]
keywords = ["llm", "ai", "deliberation", "multi-model", "council", "openrouter", "debate", "consilium"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
consilium = "consilium.cli:main"

[project.urls]
Homepage = "https://github.com/terry-li-hm/consilium"
Repository = "https://github.com/terry-li-hm/consilium"
Issues = "https://github.com/terry-li-hm/consilium/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["consilium"]
